+++ authorityspoke/io/references.py
@@ -113,6 +113,8 @@ class SelectorSchema(Schema):
 
     def split_text(self, text: str) -> Tuple[str, ...]:
         """
+        Break up shorthand text selector format into three fields.
+
         Tries to break up the string into :attr:`~TextQuoteSelector.prefix`,
         :attr:`~TextQuoteSelector.exact`,
         and :attr:`~TextQuoteSelector.suffix`, by splitting on the pipe characters.
@@ -135,7 +137,10 @@ class SelectorSchema(Schema):
         )
 
     @pre_load
-    def preprocess(self, data: Union[str, Dict[str, str]], **kwargs) -> Dict[str, str]:
+    def expand_shorthand(
+        self, data: Union[str, Dict[str, str]], **kwargs
+    ) -> Dict[str, str]:
+        """Convert input from shorthand format to normal selector format."""
         if isinstance(data, str):
             data = {"text": data}
         text = data.get("text")